safari-600.7-branch/LayoutTests/editing/pasteboard/insert-u-with-text-decoration-none.html

<!DOCTYPE html>
<html>
<body>
<div id="container"><div contenteditable>&nbsp;WebKit</div></div>
<div id="content"><u style="text-decoration: none;">hello </u><u style="text-decoration: line-through;">world</u></div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests inserting an u element with text-decoration property set to none.\n'
    + 'WebKit should not strip just font-size property. Two instances of "hello world" should look identical to each other.');

Markup.dump('container', 'Insertion point');

var content = document.getElementById('content');
Markup.dump(content, 'Content to insert');

document.querySelector('div[contenteditable]').focus();
document.execCommand('InsertHTML', null, content.innerHTML);

Markup.dump('container', 'After insertion');

</script>
</body>
</html>
