third_party/blink/web_tests/editing/pasteboard/insert-u-with-text-decoration-none-expected.txt

This tests inserting an u element with text-decoration property set to none.
WebKit should not strip just font-size property. Two instances of "hello world" should look identical to each other.

Insertion point:
| <div>
|   contenteditable=""
|   " WebKit"

Content to insert:
| <u>
|   style="text-decoration: none;"
|   "hello "
| <u>
|   style="text-decoration: line-through;"
|   "world"

After insertion:
| <div>
|   contenteditable=""
|   "hello "
|   <span>
|     style="text-decoration: line-through;"
|     "world<#selection-caret>"
|   " WebKit"
